requirements.txt
django==4.1.7
python-dotenv==1.0.0
djangorestframework==3.14.0
drf-yasg==1.21.5
django-filter==23.1
celery==5.2.7
redis==4.5.4
